Move Mac model lookup to machine checkin module and remove macmodelshelf
