+++ sal_python_pkg/sal/__init__.py
@@ -1,2 +1,1 @@
-import sal.macmodelshelf
 import sal.utils
